requirements.txt
langchain==0.1.19
langgraph==0.0.49
anthropic==0.25.1
feedparser==6.0.10
discord.py==2.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
apscheduler==3.10.4
requests==2.31.0